[#59] refactor: product 디렉토리 리팩토링
1. product > category 세부 디렉토리 생성
2. product > dto 를 세부 디렉토리 내의 dto로 이동
