+++ kyumall-client/src/main/java/com/kyumall/kyumallclient/product/category/CategoryController.java
@@ -1,8 +1,9 @@
-package com.kyumall.kyumallclient.product;
+package com.kyumall.kyumallclient.product.category;
 
-import com.kyumall.kyumallclient.product.dto.CategoryDto;
-import com.kyumall.kyumallclient.product.dto.ProductSimpleDto;
-import com.kyumall.kyumallclient.product.dto.SubCategoryDto;
+import com.kyumall.kyumallclient.product.product.ProductService;
+import com.kyumall.kyumallclient.product.category.dto.CategoryDto;
+import com.kyumall.kyumallclient.product.product.dto.ProductSimpleDto;
+import com.kyumall.kyumallclient.product.category.dto.SubCategoryDto;
 import com.kyumall.kyumallcommon.dto.CreatedIdDto;
 import com.kyumall.kyumallcommon.product.dto.CreateCategoryRequest;
 import com.kyumall.kyumallcommon.response.ResponseWrapper;
