+++ kyumall-client/src/main/java/com/kyumall/kyumallclient/product/product/ProductController.java
@@ -1,9 +1,10 @@
-package com.kyumall.kyumallclient.product;
+package com.kyumall.kyumallclient.product.product;
 
-import com.kyumall.kyumallclient.product.dto.CreateProductRequest;
-import com.kyumall.kyumallclient.product.dto.CreateProductResponse;
-import com.kyumall.kyumallclient.product.dto.ProductDetailDto;
-import com.kyumall.kyumallclient.product.dto.ProductSimpleDto;
+import com.kyumall.kyumallclient.product.StockService;
+import com.kyumall.kyumallclient.product.product.dto.CreateProductRequest;
+import com.kyumall.kyumallclient.product.product.dto.CreateProductResponse;
+import com.kyumall.kyumallclient.product.product.dto.ProductDetailDto;
+import com.kyumall.kyumallclient.product.product.dto.ProductSimpleDto;
 import com.kyumall.kyumallcommon.auth.argumentResolver.LoginUser;
 import com.kyumall.kyumallcommon.auth.authentication.AuthenticatedUser;
 import com.kyumall.kyumallcommon.response.ResponseWrapper;
